+++ nrf_802154/driver/src/nrf_802154_peripherals.h
@@ -171,6 +171,113 @@ extern "C" {
 #define NRF_802154_GPIOTE_CHANNELS_USED_MASK NRF_802154_DEBUG_GPIOTE_CHANNELS_USED_MASK
 #endif // NRF_802154_GPIOTE_CHANNELS_USED_MASK
 
+/**
+ * @def NRF_802154_EGU_NOTIFICATION_USED_CHANNELS_MASK
+ *
+ * Mask of EGU channels used by the "notification" module.
+ * See @ref NRF_802154_EGU_USED_CHANNELS_MASK.
+ */
+
+#if NRF_802154_NOTIFICATION_IMPL == NRF_802154_NOTIFICATION_IMPL_SWI
+/**
+ * @def NRF_802154_EGU_NOTIFICATION_SWI_CHANNEL_NO
+ *
+ * The channel number of the @ref NRF_802154_EGU_INSTANCE used by the
+ * "swi" implementation of the "notification" module.
+ */
+#ifndef NRF_802154_EGU_NOTIFICATION_CHANNEL_NO
+#define NRF_802154_EGU_NOTIFICATION_CHANNEL_NO 0
+#endif // NRF_802154_EGU_NOTIFICATION_SWI_CHANNEL_NO
+
+#define NRF_802154_EGU_NOTIFICATION_USED_CHANNELS_MASK \
+    (1U << NRF_802154_EGU_NOTIFICATION_CHANNEL_NO)
+#else
+#define NRF_802154_EGU_NOTIFICATION_USED_CHANNELS_MASK 0U
+#endif
+
+/**
+ * @def NRF_802154_EGU_REQUEST_USED_CHANNELS_MASK
+ *
+ * Mask of EGU channels used by the "request" module.
+ * See @ref NRF_802154_EGU_USED_CHANNELS_MASK.
+ */
+
+#if NRF_802154_REQUEST_IMPL == NRF_802154_REQUEST_IMPL_SWI
+/**
+ * @def NRF_802154_EGU_REQUEST_CHANNEL_NO
+ *
+ * The channel number of the @ref NRF_802154_EGU_INSTANCE used by the
+ * "swi" implementation of the "request" module.
+ */
+#ifndef NRF_802154_EGU_REQUEST_CHANNEL_NO
+#define NRF_802154_EGU_REQUEST_CHANNEL_NO         2
+#endif // NRF_802154_EGU_REQUEST_CHANNEL_NO
+#define NRF_802154_EGU_REQUEST_USED_CHANNELS_MASK (1U << NRF_802154_EGU_REQUEST_CHANNEL_NO)
+#else
+#define NRF_802154_EGU_REQUEST_USED_CHANNELS_MASK 0U
+#endif
+
+/**
+ * @def NRF_802154_EGU_SYNC_CHANNEL_NO
+ *
+ * The channel number of the @ref NRF_802154_EGU_INSTANCE used for generation
+ * of an interrupt by the RADIO.EVENTS_SYNC event.
+ *
+ */
+#define NRF_802154_EGU_SYNC_CHANNEL_NO            3
+
+/**
+ * @def NRF_802154_EGU_SYNC_USED_CHANNELS_MASK
+ *
+ * Mask of EGU channels used by the interrupt generation from the RADIO.EVENTS_SYNC event.
+ * See @ref NRF_802154_EGU_USED_CHANNELS_MASK.
+ */
+#define NRF_802154_EGU_SYNC_USED_CHANNELS_MASK    (1U << NRF_802154_EGU_SYNC_CHANNEL_NO)
+
+/**
+ * @def NRF_802154_EGU_RAMP_UP_CHANNEL_NO
+ *
+ * The channel number of the @ref NRF_802154_EGU_INSTANCE used for triggering the ramp-up of the RADIO.
+ */
+#define NRF_802154_EGU_RAMP_UP_CHANNEL_NO         15
+
+/**
+ * @def NRF_802154_EGU_RAMP_UP_USED_CHANNELS_MASK
+ *
+ * Mask of EGU channels used for triggering the ramp-up of the RADIO.
+ * See @ref NRF_802154_EGU_USED_CHANNELS_MASK.
+ */
+#define NRF_802154_EGU_RAMP_UP_USED_CHANNELS_MASK (1U << NRF_802154_EGU_RAMP_UP_CHANNEL_NO)
+
+/**
+ * @def NRF_802154_EGU_RAMP_UP_EVENT
+ *
+ * The EGU event used by the driver to trigger radio ramp-up.
+ */
+#define NRF_802154_EGU_RAMP_UP_EVENT              NRFX_CONCAT_2(NRF_EGU_EVENT_TRIGGERED, \
+                                                                NRF_802154_EGU_RAMP_UP_CHANNEL_NO)
+
+/**
+ * @def NRF_802154_EGU_RAMP_UP_TASK
+ *
+ * The EGU task used by the driver to trigger radio ramp-up.
+ */
+#define NRF_802154_EGU_RAMP_UP_TASK               NRFX_CONCAT_2(NRF_EGU_TASK_TRIGGER, \
+                                                                NRF_802154_EGU_RAMP_UP_CHANNEL_NO)
+
+/**
+ * @def NRF_802154_EGU_CHANNELS_MASK
+ *
+ * The mask of fixed channels of the @ref NRF_802154_EGU_INSTANCE used by the
+ * nRF 802.15.4 Radio Driver.
+ */
+#define NRF_802154_EGU_USED_CHANNELS_MASK             \
+    (NRF_802154_EGU_NOTIFICATION_USED_CHANNELS_MASK | \
+     NRF_802154_EGU_REQUEST_USED_CHANNELS_MASK |      \
+     NRF_802154_EGU_SYNC_USED_CHANNELS_MASK |         \
+     NRF_802154_EGU_RAMP_UP_USED_CHANNELS_MASK |      \
+     NRF_802154_SL_EGU_USED_CHANNELS_MASK)
+
 #ifdef __cplusplus
 }
 #endif